Combine files and simplify file names
- circlehash64.go is for Go 1.17 and newer versions. It
uses some optimizations that make the code less readable.

- circlehash64_ref.go is for Go 1.16 and older versions. It
is the reference implementation that is easy to read and audit.

NOTE: CI using GitHub Actions will test both implementations.

CAVEATS: There is some redundancy between both implementations
but that's intentional.  The functions in circlehash64.go
can be optimized, making currently identical functions different
from circlehash64_ref.go while remaining 100% compatible.
